add config instructions for ember-try
Nowadays addons are initialized with `ember-try` configured to test older version of ember by default. These older versions are not compatible with the current `ember-font-awesome` version, so I added some guidance for correctly configuring `ember-try` to it install and use the proper version of `ember-font-awesome` when testing older versions of ember.